Moved controller handler functions to sperate files
