+++ src/controllers/timeline.go
@@ -1,0 +1,84 @@
+package controllers
+
+import (
+	"crypto/md5"
+	"encoding/hex"
+	"github.com/gin-gonic/gin"
+	"github.com/noirbizarre/gonja"
+	"minitwit/src/database"
+	"minitwit/src/functions"
+	"minitwit/src/models"
+	"net/http"
+	"strconv"
+	"strings"
+)
+
+var timelineTemplate = gonja.Must(gonja.FromFile("templates/timeline.html"))
+
+var g models.Session
+
+func HandleUserTimeline(w http.ResponseWriter, r *http.Request, c *gin.Context, username string) {
+
+	user := database.GetUserFromDb(username)
+
+	messages := database.GetUserMessages(user.User_id)
+	twits := convertMessagesToTwits(&messages)
+	out, err := timelineTemplate.Execute(gonja.Context{"g": g, "request": r, "messages": twits})
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+	w.Write([]byte(out))
+}
+
+func HandleTimeline(w http.ResponseWriter, r *http.Request, c *gin.Context) {
+	// Execute the template per HTTP request
+	request := functions.GetEndpoint(r)
+	data, err := functions.GetCookie(c)
+	g = data
+
+	// If there is no cookie
+	if err != nil || g.User.Username == "" {
+		c.Redirect(http.StatusFound, "/public")
+	}
+
+	//set g = "None" if g.user should return false in jinja
+
+	out, err := timelineTemplate.Execute(gonja.Context{"g": g, "request": request})
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+	w.Write([]byte(out))
+}
+
+func HandlePublicTimeline(w gin.ResponseWriter, r *http.Request, c *gin.Context) {
+	request := functions.GetEndpoint(r)
+
+	messages := database.GetAllMessages()
+
+	twits := convertMessagesToTwits(&messages)
+	//print(string(request))
+	out, err := timelineTemplate.Execute(gonja.Context{"g": g, "request": request, "messages": twits})
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+	w.Write([]byte(out))
+}
+
+func convertMessagesToTwits(messages *[]models.Message) []models.Twit {
+	var twits []models.Twit
+	for _, message := range *messages {
+		twits = append(twits, models.Twit{getGavaterUrl(message.Email, 48), message.Username, strconv.Itoa(int(message.Pubdate)), message.Text})
+	}
+	print(twits)
+	return twits
+
+}
+
+func getGavaterUrl(email string, size int) string {
+	data := []byte(strings.ToLower(strings.TrimSpace(email)))
+	hash := md5.Sum(data)
+	hashStr := hex.EncodeToString(hash[:])
+
+	str := []string{"http://www.gravatar.com/avatar/", hashStr, "?d=identicon&s=", strconv.Itoa(size)}
+	return strings.Join(str, "")
+}
